requirements.txt
qdrant-client==1.9.1
python-dotenv==1.0.1
matplotlib==3.9.0
ipython==8.25.0
openai==1.35.7
numpy==1.26.4
openai-whisper
# pytube==15.0.0
pytubefix==6.8.1
moviepy==1.0.3
langchain-community==0.2.11
langchain-ollama==0.1.1
unstructured==0.15.0
streamlit==1.37.0